merged branch for game over, needs test

+++ peripherals/include/galaga_bitmaps.h
@@ -73,7 +73,7 @@ void lcd_draw_explosion(
 );
 
 void itoa(
-	uint16_t source,
+	uint32_t source,
 	char* dest_ADDRESS_OF_SIZE_NINE_CHAR_ARRAY_PLEASE_AND_THANK_YOU
 );
 #endif
